pyproject.toml
[project]
name = "bible-rag-agent-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi>=0.116.1",
    "groq>=0.30.0",
    "jinja2>=3.1.6",
    "numpy==1.26.0",
    "pgvector>=0.4.1",
    "psycopg2-binary>=2.9.10",
    "pydantic-ai>=0.4.4",
    "pydantic-ai-slim[groq]>=0.4.4",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.4",
    "rich>=14.0.0",
    "sentence-transformers>=5.0.0",
    "sqlalchemy>=2.0.41",
    "starlette>=0.47.2",
    "torch==2.2.2",
    "transformers>=4.53.2",
    "uvicorn[standard]>=0.35.0",
]
